Changed the way of running Parallax V2 (use -v2 args).
Usage: python ./run-parallax.py -v2
